feat(backend): Moved public and private key reading function into config Separated registration and authentication routes into two separate routes. Also added validation function.Addressed review comments.
