+++ backend/src/config/Config.ts
@@ -1,10 +1,41 @@
-import * as dotenv from "dotenv";
-
-dotenv.config({ path: "backend/**/*.env" });
+import * as fs from 'fs';
+import * as path from 'path';
 
 class Config {
   dbHost: string;
+  privateKeyPath: string;
+  publicKeyPath: string;
+
+  constructor(privateKeyPath: string, publicKeyPath: string) {
+    this.dbHost = process.env.DB_HOST || '';
+    this.privateKeyPath = privateKeyPath;
+    this.publicKeyPath = publicKeyPath;
+  }
+
+  readPrivateKey(): string {
+    try {
+      return fs.readFileSync(this.privateKeyPath, 'utf8');
+    } catch (err) {
+      console.error("Ошибка чтения приватного ключа:", err);
+      throw new Error('Не удалось прочитать файл приватного ключа');
+    }
+  }
+
+  readPublicKey(): string {
+    try {
+      return fs.readFileSync(this.publicKeyPath, 'utf8');
+    } catch (err) {
+      console.error("Ошибка чтения публичного ключа:", err);
+      throw new Error('Не удалось прочитать файл публичного ключа');
+    }
+  }
 }
 
-export const config = new Config();
-config.dbHost = process.env.DB_HOST;
+// Путь к файлам ключей
+const privateKeyPath = path.resolve('src/private.pem');
+const publicKeyPath = path.resolve('src/public.pem');
+
+export const config = new Config(
+  privateKeyPath,
+  publicKeyPath
+);
